+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [#811]: `book`: Add some examples for byte slice/array hints as well
 - [#800]: `defmt-macros`: Fix generic trait bounds in Format derive macro
 
-[#821]: https://github.com/knurling-rs/defmt/pull/xxx
+[#821]: https://github.com/knurling-rs/defmt/pull/821
 [#813]: https://github.com/knurling-rs/defmt/pull/813
 [#811]: https://github.com/knurling-rs/defmt/pull/811
 [#800]: https://github.com/knurling-rs/defmt/pull/800